Add support for importing by <ORG NAME>/<WORKSPACE NAME>
This changes the Importer function for the Workspace resource to handle passing an
organization name and workspace name pair to import the workspace. If no pair is
specified the importer function will assume an opaque external id was passed.

An example import command looks as follows:

terraform import tfe_workspace.test my-org-name/my-wkspace-name